Chat Room
This chat service will contain a main lobby where users sign on with a
nickname and can communicate with each other. Users may also create
chat rooms for other to join. The entire app should be displayed on a
single webpage, listing the room you are in, all available rooms, and
the users in the current room.